Add remaining tests (#11); add small fixes to prm to facilitate testing; version bump
Add test of project script existence in project start and stop
Small fixes to prm to facilitate testing
Make test-variable available during tests
Add if-tests to prm to avoid hanging on $EDITOR file-edits in tests
Add remaining unit tests

+++ prm.sh
@@ -3,7 +3,7 @@
 # Copyright (c) 2015 Eivind Arvesen. All Rights Reserved.
 
 COPY="Written by Eivind Arvesen, 2015."
-VERSION=0.4.0
+VERSION=0.5.0
 SOURCE=$(printf "prm MUST be sourced - not run in a subshell.\ni.e. '. ./prm'\n" && echo " ")
 
 function return_error() {
@@ -153,18 +153,22 @@ case "$1" in
             for argument in "${@:2}"; do
                 if [ -d "$prm_dir/$argument" ]; then
                     return_error 1 "Project $argument already exists"
+                    return 1
                 else
-                    check_project_name "$2" || return
+                    check_project_name "$argument" || return
                     check_editor || return
                     mkdir -p "$prm_dir/$argument"
                     printf "#!/usr/bin/env bash\n\n# This script will run when STARTING the project \"%s\"\n# Here you might want to cd into your project directory, activate virtualenvs, etc.\n\n# The currently active project is available via \$PRM_ACTIVE_PROJECT\n# Command line arguments can be used, \$3 would be the first argument after your project name.\n\n" "$argument" > "$prm_dir/$argument/start.sh"
                     printf "#!/usr/bin/env bash\n\n# This script will run when STOPPING the project \"%s\"\n# Here you might want to deactivate virtualenvs, clean up temporary files, etc.\n\n# The currently active project is available via \$PRM_ACTIVE_PROJECT\n# Command line arguments can be used, \$3 would be the first argument after your project name.\n\n" "$argument" > "$prm_dir/$argument/stop.sh"
-                    $EDITOR "$prm_dir/$argument/start.sh" && $EDITOR "$prm_dir/$argument/stop.sh"
+                    if [ ! $prm_bats_tests ]; then
+                        $EDITOR "$prm_dir/$argument/start.sh" && $EDITOR "$prm_dir/$argument/stop.sh"
+                    fi
                     echo "Added project $argument"
                 fi
             done
         else
             return_error 1 "No name given"
+            return 1
         fi
         ;;
     copy)
@@ -172,24 +176,30 @@ case "$1" in
         if [ "$2" ]; then
             if [ ! -d "$prm_dir/$2" ]; then
                 return_error 1 "$2: No such project"
+                return 1
             else
                 if [ "$3" ]; then
                     if [ -d "$prm_dir/$3" ]; then
                         return_error 1 "Project $3 already exists"
+                        return 1
                     else
                         check_project_name "$3" || return
                         check_editor || return
                         cp -r "$prm_dir/$2" "$prm_dir/$3"
                         sed -i -e "s/\"$2\"/\"$3\"/g" $prm_dir/$3/*.sh
-                        $EDITOR "$prm_dir/$argument/start.sh" && $EDITOR "$prm_dir/$argument/stop.sh"
+                        if [ ! $prm_bats_tests ]; then
+                            $EDITOR "$prm_dir/$argument/start.sh" && $EDITOR "$prm_dir/$argument/stop.sh"
+                        fi
                         echo "Copied project $2 to $3"
                     fi
                 else
                     return_error 1 "No new name given"
+                    return 1
                 fi
             fi
         else
             return_error 1 "No name given"
+            return 1
         fi
         ;;
     edit)
@@ -198,20 +208,25 @@ case "$1" in
             for argument in "${@:2}"; do
                 if [ -d "$prm_dir/$argument" ]; then
                     check_editor || return
-                    $EDITOR "$prm_dir/$argument/start.sh" && $EDITOR "$prm_dir/$argument/stop.sh"
+                    if [ ! $prm_bats_tests ]; then
+                        $EDITOR "$prm_dir/$argument/start.sh" && $EDITOR "$prm_dir/$argument/stop.sh"
+                    fi
                     echo "Edited project $argument"
                 else
                     return_error 1 "$argument: No such project"
+                    return 1
                 fi
             done
         else
             return_error 1 "No name given"
+            return 1
         fi
         ;;
     list)
         # List projects
         if [ ! "$(find "$prm_dir" -type d | wc -l)" -gt 2 ]; then
             return_error 1 "No projects exist"
+            return 1
         else
             cd "$prm_dir/" >/dev/null 2>&1 || return_error 1 "Directory $prm_dir does not exist."
             for active in ./*; do
@@ -226,41 +241,49 @@ case "$1" in
             for argument in "${@:2}"; do
                 if [ -e "$prm_dir/.active-$$.tmp" ] && [ "$(cat "$prm_dir/.active-$$.tmp")" == "$argument" ]; then
                     return_error 1 "Stop project $argument before trying to remove it"
+                    return 1
                 else
                     if [ -d "$prm_dir/$argument" ]; then
                         rm -rf "${prm_dir:?}/$argument/"
                         echo "Removed project $argument"
                     else
                         return_error 1 "$argument: No such project"
+                        return 1
                     fi
                 fi
             done
         else
             return_error 1 "No name given"
+            return 1
         fi
         ;;
     rename)
         # Rename project
         if [ -e "$prm_dir/.active-$$.tmp" ] && [ "$(cat "$prm_dir/.active-$$.tmp")" == "$2" ]; then
             return_error 1 "Stop project $2 before trying to rename it"
+            return 1
         else
             if [ "$2" ]; then
                 if [ ! -d "$prm_dir/$2" ]; then
                     return_error 1 "$2: No such project"
+                    return 1
                 else
                     if [ "$3" ]; then
                         if [ -d "$prm_dir/$3" ]; then
                             return_error 1 "Project $3 already exists"
+                            return 1
                         else
                             mv "$prm_dir/$2" "$prm_dir/$3"
                             echo "Renamed project $2 $3"
                         fi
                     else
                         return_error 1 "No new name given"
+                        return 1
                     fi
                 fi
             else
                 return_error 1 "No name given"
+                return 1
             fi
         fi
         ;;
@@ -270,6 +293,7 @@ case "$1" in
             if [ -d "$prm_dir/$2" ]; then
                 if [ -e "$prm_dir/.active-$$.tmp" ] && [ "$(cat "$prm_dir/.active-$$.tmp")" == "$2" ]; then
                     return_error 1 "Project $2 is already active"
+                    return 1
                 else
                     if [ ! -e "$prm_dir/.path-$$.tmp" ]; then
                         pwd > "$prm_dir/.path-$$.tmp"
@@ -278,23 +302,30 @@ case "$1" in
                         . "$prm_dir/$(cat "$prm_dir/.active-$$.tmp")/stop.sh"
                         PRM_ACTIVE_PROJECT=""
                     fi
-                    echo "$2" > "$prm_dir/.active-$$.tmp"
-                    set_prompt_start "$2"
-                    echo "Starting project $2"
-                    . "$prm_dir/$2/start.sh"
-                    PRM_ACTIVE_PROJECT=$2
+                    if [ -e "$prm_dir/$2/start.sh" ] && [ -e "$prm_dir/$2/stop.sh" ]; then
+                        echo "$2" > "$prm_dir/.active-$$.tmp"
+                        set_prompt_start "$2"
+                        echo "Starting project $2"
+                        . "$prm_dir/$2/start.sh"
+                        PRM_ACTIVE_PROJECT=$2
+                    else
+                        return_error 1 "Cannot start project $2: Project has no scripts"
+                        return 1
+                    fi
                 fi
             else
                 return_error 1 "$2: No such project"
+                return 1
             fi
         else
             return_error 1 "No name given"
+            return 1
         fi
         ;;
     stop)
         # Stop project
         if [ -e "$prm_dir/.active-$$.tmp" ]; then
-            . "$prm_dir/$(cat "$prm_dir/.active-$$.tmp")/stop.sh"
+            . "$prm_dir/$(cat "$prm_dir/.active-$$.tmp")/stop.sh" || return_error 1 "Cannot stop project $project_name: Project has no stop script"
             echo "Stopping project $(cat "$prm_dir/.active-$$.tmp")"
             rm -f "$prm_dir/.active-$$.tmp"
             cd "$(cat "$prm_dir/.path-$$.tmp")" >/dev/null 2>&1 || return_error 1 "Could not change directory to original path."
@@ -304,6 +335,7 @@ case "$1" in
             PRM_ACTIVE_PROJECT=""
         else
             return_error 1 "No active project"
+            return 1
         fi
         ;;
     -h|--help)
